--- dms-backend/app/core/security.py
@@ -1,26 +1,0 @@
-from datetime import datetime
-from datetime import timedelta
-from typing import Optional
-from app.core.config import settings
-from jose import jwt
-
-
-def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.utcnow() + expires_delta
-    else:
-        expire = datetime.utcnow() + timedelta(
-            minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES
-        )
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(
-        to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM
-    )
-    return encoded_jwt
-
-
-def decode_access_token(access_token: str):
-    decoded_jwt = jwt.decode(access_token,
-                             settings.SECRET_KEY, settings.ALGORITHM)
-    return decoded_jwt['sub']
