+++ dms-backend/app/db/base.py
@@ -1,2 +1,15 @@
-from app.db.base_class import Base
-from app.db.models.user import User
+from typing import Any
+
+from sqlalchemy.ext.declarative import declared_attr
+from sqlalchemy.orm import as_declarative
+
+
+@as_declarative()
+class Base:
+    id: Any
+    __name__: str
+
+    # to generate tablename from classname
+    @declared_attr
+    def __tablename__(cls) -> str:
+        return cls.__name__.lower()
